move lsr test to celestial transformations file
